+++ changelog.md
@@ -6,6 +6,7 @@
 * Added Smartness enchantment
   * Increases experience dropped by mobs by 50% per level. Incompatible with Looting
 * Added durability to gear and increased durability of some materials
+* Tiredness is no longer reset on death
 * Fixes and advancements
 
 ## Alpha 3.9.1